config: add separate config for LKMs
Add a separate configuration file for Linux loadable kernel modules (LKMs). Use
this one instead of the default configuration when analyzing an LKM and install
it to the same location as the default configuration. Limit the checks
that are executed to those that make sense in kernel space.

Cases where an LKM is analyzed are recognized by checking if a
relocatable object file contains the ".modinfo" and
".gnu.linkonce.this_module" sections.

Multiple checks, e.g., CWE789, CWE416, and CWE134, require more work to
function properly for LKMs. We mainly need to teach the cew_checker
about the parameters and semantics of some functions. They are included,
but effectively disabled since their configuration is empty. CWE467
actually works since the functions are called the same as in user space
and thus Ghidra recognizes their parameters.