ir: impl `Display` for `Program`
Implement the `Display` trait for the intermediate language
representation of the analyzed program.

This outputs a human-friendly representation that is more compact than
what is returned by `Debug` and is therefore easier to consume.

There are no direct users of this implementation, its main purpose is
to facilitate easier debugging during development.