lib/analysis/taint: introduce `handle_empty_state_out` callback
Add a callback that allows a taint analysis to hook into the fixpoint
computation when some transfer function maps its input state(s) to the
empty state.

For some analyses this event may be a sink, e.g., for CWE252, while for
many analyses it does not make sense to propagate empty states further
since it is impossible to generate a non-empty state from them; they
may use this hook to optimize resource usage.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>