lib/utils/debug: introduce `ToJsonCompact` trait
Many types implement a custom JSON serialization method for
internal debugging purposes.

Add an abstraction for this pattern in form of the `ToJsonCompact` trait.
This enables all types to benefit from the default implementation of a
printing method and makes it easier to use generic programming.

This commit does not convert any existing types that implement this
behavior. They are expected to be converted in an ongoing process.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>