lib/analysis/taint: propagate memory taint
Remove the limitation that `update_return_callee` can not be used to
propagate memory taint from the callee to the caller. Do so by renaming
abstract identifiers in the default implementation of `update_return` in
the `TaintAnalysis` trait. In particular, implementers of
`update_return_callee` do not have to case about renaming and can return
the abstract identifiers of the callee context.

Adjust CWE252 to make use of this new feature.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>