lib/abstract_domain: add `merge_with` method to `AbstractDomain`
The `merge` method always produces a new, owned value. This might
be undesirable in situations where it is possible to modify an
existing value in-place.

Add a new method that allows an abstract domain to provide a method to
merge one object into another in-place.

It is a common pattern to see something like this:

```
*mut_ref = mut_ref.merge(other_ref);
```

where `mut_ref` is a mutable reference to a type that implements
`AbstractDomain`. Note that it is common that such types are just
wrappers around an `Arc` to an inner type that is expensive to clone.
However, while cloning of one of the refs in `merge` may be cheap, it
means that then there are >= 2 references to the underlying `Arc`, which
means that it can never be cheaply modified, i.e., the retuned owned
value will usually involve an expensive clone. However:

```
mut_ref.merge_with(other_ref);
```

can potentially do a cheap modification of the underlying `Arc`.

Due to the default implementation it should always be OK to replace the
first pattern with the second in generic code, i.e., it never decreases
performance and can only increase it if the type provides an optimized
implementation.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>