lib/abstract_domain/domain_map: add `merge_map_with` to `MapMergeStrategy`, implement `merge_with`
Add a `merge_map_with` method to the `MapMergeStrategy` and provide a
default implementation in terms of it for the `merge_map` method.

Use `merge_map_with` to provide an optimized implementation of
`merge_with` for `DomainMap`.

Convert all existing implementations of `MapMergeStrategy` to implement
`merge_map_with` instead.

The rationale is change similar to the one detailed in Commit("
7c0ffbe lib/abstract_domain: add `merge_with` method to
`AbstractDomain`").

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>