lib/analysis/taint: use `DomainMap` for register and memory taint
In the original implementation of memory taint propagation in
Commit("e7c25f7 lib/analysis/taint: propagate memory taint") it was
overlooked that we already have an abstraction for maps into abstract
domains that are abstract domains themselves.

Use the `DomainMap` abstraction for the register and memory taint maps
that make up the `State` type.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>