lib/analysis/taint/state: overwrite memory taint in more cases
Currently we only overwrite memory taint if the PI result for the
target address is very exact.

Weaken the conditions under which we overwrite taint information by
allowing possibly constant or top values for the target address
as long as the target memory object and offset are unique.

This may lead to taint being overwritten with non-tainted values in more
cases.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>