lib/checkers/cwe476: update to use new TA features
By removing the `update_return_callee` implementation we get the default
implementation, which implies that we now do an interprocedural
analysis. This means we no longer raise a waning when a function may
return a NULL pointer, but rather make the caller responsible for
checking it.

Introduce a new configuration parameter `strict_call_policy` that can be
used to enable the current default behavior, where any taint that may be
passed as an argument to a function call is causing a warning. This is
to prepare for the introduction of a new mode, where a more fine grained
policy is used.

Remove the `extern_symbol_map` and `current_sub` members from the
`Context` type. Those are no longer needed in this version of the check.

Include a 'reason' field in the generated CWE warnings s.t. it is easier to
check which heuristic triggered the warning.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>