lib/checkers/cwe476: use function signatures for in-binary calls
Currently, we only emit warnings for function calls if `strict_call_policy`
is set. However, enabling this can cause many FPs since warnings are
emitted as soon as any parameter may contain taint, or point to a tainted,
value. For nested parameters it may also lead to FNs since only one level
of nesting is considered.

Use function summaries to make a more accurate decision whether or not a
callee may dereference a potential NULL pointer that is made available to
them.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>