lib/ir/project: add pass to retarget returns from non-returning calls
The original fix for Issue #461 in Commit ("lib/ir/project: propagate
control flow for call returns") was incomplete.

The original problem was due to a call to a function without a return
instruction "returning" to a block that could be optimized away in the
propagate control flow pass. Retargeting the call return can only solve
the issue when the return block can be retargeted (and the retarget is
not optimized away), which is not the case for condition blocks.

Thus, always retarget returns from calls to functions without a ret
to the artificial sink.

Link: https://github.com/fkie-cad/cwe_checker/pull/462#issuecomment-2093656674
Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>