lib/pcode: fix zero-sized subpiece due to bad subregister substitution
Assignments to subregisters that have the same name as the full register
may lead to expressions that have SUBPIECE operations that would result
in zero-sized values. In general, the subregister substitution produces
semantically incorrect IR code in those cases.

I observed that this occurs when assignments are made to the lower
bytes of a larger register. Those cases should be handled by this
patch. If this may also happen for assignments to other subregisters,
then this patch is incomplete (see FIXME).

Reported-by: https://github.com/ljungnickel
Closes: https://github.com/fkie-cad/cwe_checker/issues/469
Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>