lib: explicitly opt-in to overflowing arithmetic
Replace integer arithmetic that might overflow with the appropriate
methods to explicitly opt-in to the behavior.

Since these methods have the same semantics as the bare operators (in
release builds) there should be no semantic changes here.

I did not investigate if the original code was aware that these
operations might overflow. In particual, it is not clear to me if the
right shift is expected to potentially overflow.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>