lib/utils/debug: introduce infrastructure for per-stage debugging
While debugging analysis issues I found myself reinserting the
same print statements at the same places for each new investigation.

Therefore, introduce an extensible infrastructure to debug different
parts of the analysis. The `utils::debug::Settings` object is derived
from the command line and an immutable reference to it is the sole thing
that is being passed around to convey the current debug environment to
different part of the code. It also provides helper methods to perform
debug operations based on the runtime environment.

The related command-line options and the output format are entirely
unstable and only meant for development purposes.

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>