lib/utils/debug: add option to read output of Ghidra plugin from file
The output of the Ghidra plugin (and Ghidra in general) is
non-deterministic, even on the same Ghidra version. Probably this is
due to analysis timeouts.

Furthermore, running Ghidra takes a significant portion of the analysis
time.

Therefore, we can create reproducibility as well as speedups during
development by saving the plugin output (using `--debug pcode-raw`) and
replaying it later (via `--pcode-raw /path/to/file`).

Signed-off-by: Valentin Obst <valentin.obst@fkie.fraunhofer.de>